refactor code layout for readability
