+++ aggregate.go
@@ -1,0 +1,85 @@
+package main
+
+import (
+	"io"
+	"log"
+	"net/http"
+	"sort"
+	"sync"
+
+	dto "github.com/prometheus/client_model/go"
+	"github.com/prometheus/common/expfmt"
+)
+
+type aggregate struct {
+	familiesLock sync.RWMutex
+	families     map[string]*dto.MetricFamily
+}
+
+func newAggregate() *aggregate {
+	return &aggregate{
+		families: map[string]*dto.MetricFamily{},
+	}
+}
+
+func (a *aggregate) parseAndMerge(r io.Reader) error {
+	var parser expfmt.TextParser
+	inFamilies, err := parser.TextToMetricFamilies(r)
+	if err != nil {
+		return err
+	}
+
+	a.familiesLock.Lock()
+	defer a.familiesLock.Unlock()
+	for name, family := range inFamilies {
+		// Sort labels in case source sends them inconsistently
+		for _, m := range family.Metric {
+			sort.Sort(byName(m.Label))
+		}
+
+		if err := validateFamily(family); err != nil {
+			return err
+		}
+
+		// family must be sorted for the merge
+		sort.Sort(byLabel(family.Metric))
+
+		existingFamily, ok := a.families[name]
+		if !ok {
+			a.families[name] = family
+			continue
+		}
+
+		merged, err := mergeFamily(existingFamily, family)
+		if err != nil {
+			return err
+		}
+
+		a.families[name] = merged
+	}
+
+	return nil
+}
+
+func (a *aggregate) handler(w http.ResponseWriter, r *http.Request) {
+	contentType := expfmt.Negotiate(r.Header)
+	w.Header().Set("Content-Type", string(contentType))
+	enc := expfmt.NewEncoder(w, contentType)
+
+	a.familiesLock.RLock()
+	defer a.familiesLock.RUnlock()
+	metricNames := []string{}
+	for name := range a.families {
+		metricNames = append(metricNames, name)
+	}
+	sort.Sort(sort.StringSlice(metricNames))
+
+	for _, name := range metricNames {
+		if err := enc.Encode(a.families[name]); err != nil {
+			log.Println("An error has occurred during metrics encoding:\n\n" + err.Error())
+			return
+		}
+	}
+
+	// TODO reset gauges
+}
